Add `in` and `not_in` rules.
These were in the documentation on CoderWall, but were not actually present
in the code.

Since I needed them, and I suspect others do too, I decided to go ahead and
implement them based on the existing logic in the `helper.go` file, and on
the other rules.

I've added test cases for success and failure, for `in` and `not_in`, and
in both cases for where the range is either an integer or a string. I know
that the values get converted to strings anyway, but I was trying to be
extra thorough :)